Add transaction (multi/exec/discard) support.
Turns out everything has a potential return type of <<"queued">> when
using transactions.  Ugly, but doable.  We may want to make the return
functions nested so we could have the <<"queued">> check in one place.

The give-me-a-fun-and-i-do-you-a-txn style is inspired by mnesia.  Just
give er:er_transaction/2 or erp:er_transaction/1 a one-airity fun, use
the parameter to the fun as your connection back to redis, and
everything you need will Just Work.